slightly more idiomatic specs
These changes mainly consolidate the spec blocks (`it`) down to the expectations
while moving the other parts of the tests into `let` and `subject` blocks.

A few things I think are super cool:

1. Using RSpec::Matchers#change like so:

    subject(:sync) { -> { Note.sync(attributes) } }

    context 'when the guid is new' do
      let(:attributes) { build(:note).attributes }
      it { should change(Note, :count).from(0).to(1) }
    end

defines the subject (accessible as `sync`) as a lambda which calls `Note.sync`.
Since the subject is a lamba we can say `it { should change(...) }` which lets
us compare a value before and after the lambda is called.  In this case, it is
checking `Note.count` before and after the lambda is called.

2. Subtle `let!`:

    let!(:note) { create(:note, en_updated_at: 2.days.ago) }

this both creates the `note` variable and initializes it before our tests are
run.  This is equivalent to doing:

    let(:note) { create(:note, en_updated_at: 2.days.ago) }
    before { note }